* bleeding-jumbo: (31 commits)
  Update to hi res timer code.
  Add some OSX building tips to docs.
  Don't keep salt hash table if --save-memory=2 or greater, or if the format doesn't have a salt_hash function anyway.  Update salt hash table after sorting (if applicable).  Add some safety measures to pot reload.
  The documentation of --save-memory was duplicated in doc/OPTIONS.
  Keep loader's salt hash table for pot reload (updating it w/ required).
  Less timid --max-runtime: We now give a 3 sec grace period and then force quit as if ctrl-c was pressed twice.
  "Reload" code cleanup and optimizations.
  #ifdef out some select() stuff in tty_getchar() that was used for Cygwin but isn't needed (anymore?).
  Speed up pot reload (a lot) by using bitmap.
  Recent structural changes and new features introduced some terrific bugs. This hopefully fixes most of them.
  Pause/abort on filesystem stimuli. This defaults to /var/run/john/abort and /var/run/john/pause but can be reconfigured to any other files.
  Drop an unneeded strnzcpy() during reload of pot file. Also adds debug stuff to crk_reload_pot().
  Do not re-read pot file lines that were processed by initial loading.
  Drop the experimental --reload-every=N option. We always reload at session save (unless totally disabled in john.conf).
  MPI: Add signaling of cracked hashes. There is also various changes to the handling of --fork sessions in this regard.
  Various signal.c tweaks for MPI: Avoid signal loops when signaling parent for reload. If running Cygwin w/ fork or MPI, save session on key press in case a following termination does not save properly. Also clarify MPI requirements in doc/README-mpi.
  vncpcap2john: remove bogus NUL check in makehex()
  Makefile: add proper dependency information for vncpcap2john
  vncpcap2john: add license blurb for own contribs
  fix Makefile for vncpcap2john
  ...

Conflicts:
	src/john-mpi.c
	src/loader.c
	src/logger.c
	src/vncpcap2john.cpp

+++ src/signals.c
@@ -69,10 +69,12 @@
 volatile int event_pending = 0, event_reload = 0;
 volatile int event_abort = 0, event_save = 0, event_status = 0;
 volatile int event_ticksafety = 0;
+volatile int event_mpiprobe = 0, event_poll_files = 0;
 
-volatile int timer_reload = 0, timer_abort = 0, timer_status = 0;
+volatile int timer_abort = 0, timer_status = 0;
 static int timer_save_interval, timer_save_value;
 static clock_t timer_ticksafety_interval, timer_ticksafety_value;
+volatile int aborted_by_timer = 0;
 
 #if !OS_TIMER
 
@@ -120,28 +122,26 @@ void sig_timer_emu_tick(void)
 
 #endif
 
-static void sig_install_update(void);
+static void sig_install(void *handler, int signum)
+{
+#ifdef SA_RESTART
+	struct sigaction sa;
+
+	memset(&sa, 0, sizeof(sa));
+	sa.sa_handler = handler;
+	sa.sa_flags = SA_RESTART;
+	sigaction(signum, &sa, NULL);
+#else
+	signal(signum, handler);
+#endif
+}
 
 static void sig_handle_update(int signum)
 {
 	event_reload = event_save = event_pending = 1;
 
 #ifndef SA_RESTART
-	sig_install_update();
-#endif
-}
-
-static void sig_install_update(void)
-{
-#ifdef SA_RESTART
-	struct sigaction sa;
-
-	memset(&sa, 0, sizeof(sa));
-	sa.sa_handler = sig_handle_update;
-	sa.sa_flags = SA_RESTART;
-	sigaction(SIGHUP, &sa, NULL);
-#else
-	signal(SIGHUP, sig_handle_update);
+	sig_install(sig_handle_update, signum);
 #endif
 }
 
@@ -157,23 +157,19 @@ void check_abort(int be_async_signal_safe)
 	tty_done();
 
 	if (be_async_signal_safe) {
-		if (timer_abort >= 0) {
-			if (john_main_process)
+		if (john_main_process) {
+			if (aborted_by_timer)
+				write_loop(2, "Session stopped (max run-time"
+				           " reached)\n", 39);
+			else
 				write_loop(2, "Session aborted\n", 16);
-#if defined(HAVE_MPI) && defined(JOHN_MPI_ABORT)
-			if (mpi_p > 1)
-				MPI_Abort(MPI_COMM_WORLD,1);
-#endif
-		} else
-		if (john_main_process)
-			write_loop(2, "Session stopped (max run-time"
-			           " reached)\n", 39);
+		}
 		_exit(1);
 	}
 
 	if (john_main_process)
-		fprintf(stderr, "Session %s\n", (timer_abort >= 0) ?
-		        "aborted" : "stopped (max run-time reached)");
+		fprintf(stderr, "Session %s\n", (aborted_by_timer) ?
+		        "stopped (max run-time reached)" : "aborted");
 	error();
 }
 
@@ -215,9 +211,7 @@ static void sig_install_abort(void)
 #endif
 
 	signal(SIGINT, sig_handle_abort);
-//#ifndef HAVE_MPI
 	signal(SIGTERM, sig_handle_abort);
-//#endif
 #ifdef SIGXCPU
 	signal(SIGXCPU, sig_handle_abort);
 #endif
@@ -233,9 +227,7 @@ static void sig_remove_abort(void)
 #endif
 
 	signal(SIGINT, SIG_DFL);
-//#ifndef HAVE_MPI
 	signal(SIGTERM, SIG_DFL);
-//#endif
 #ifdef SIGXCPU
 	signal(SIGXCPU, SIG_DFL);
 #endif
@@ -255,49 +247,53 @@ static void signal_children(int signum)
 #endif
 
 static void sig_install_timer(void);
+static void sig_handle_reload(int signum);
 
 static void sig_handle_timer(int signum)
 {
 	int saved_errno = errno;
 #ifndef BENCH_BUILD
-
+	/* Some stuff only done every third second */
+	if ((timer_save_value & 3) == 3) {
+#ifdef HAVE_MPI
+		if (!event_reload && mpi_p > 1) {
+			event_pending = event_mpiprobe = 1;
+		}
+#endif
+		event_poll_files = event_pending = 1;
+		sig_install(sig_handle_reload, SIGUSR2);
+	}
 #if OS_TIMER
 	if (!--timer_save_value) {
 		timer_save_value = timer_save_interval;
-		event_reload = event_save = event_pending = 1;
+		event_save = event_pending = 1;
+		event_reload = options.reload_at_save;
 	}
-
 	if (timer_abort && !--timer_abort) {
-		timer_abort = -1;
-		event_abort = event_pending = 1;
+		aborted_by_timer = 1;
+		timer_abort = 3;
+		sig_handle_abort(0);
 	}
 	if (timer_status && !--timer_status) {
 		timer_status = options.status_interval;
 		event_status = event_pending = 1;
-	}
-	if (timer_reload && !--timer_reload) {
-		timer_reload = options.reload_interval;
-		event_reload = 1;
 	}
 #else /* no OS_TIMER */
 	unsigned int time = status_get_time();
 
 	if (time >= timer_save_value) {
 		timer_save_value += timer_save_interval;
-		event_reload = event_save = event_pending = 1;
+		event_save = event_pending = 1;
+		event_reload = options.reload_at_save;
 	}
-
 	if (timer_abort && time >= timer_abort) {
-		timer_abort = -1;
-		event_abort = event_pending = 1;
+		aborted_by_timer = 1;
+		timer_abort += 3;
+		sig_handle_abort(0);
 	}
 	if (timer_status && time >= timer_status) {
 		timer_status += options.status_interval;
 		event_status = event_pending = 1;
-	}
-	if (timer_reload && time >= timer_reload) {
-		timer_reload += options.reload_interval;
-		event_reload = 1;
 	}
 #endif /* OS_TIMER */
 #endif /* !BENCH_BUILD */
@@ -398,34 +394,42 @@ static void sig_remove_timer(void)
 
 static void sig_handle_status(int signum)
 {
-	/* For cygwin and MPI, we save too here */
-	//event_save =
+	/* We currently disable --fork for Cygwin in os.h due to problems
+	   with proper session save when a job is aborted. This cludge
+	   could be a workaround: First press a key, then abort it after
+	   status line was printed. */
+#if OS_FORK && defined(__CYGWIN32__)
+	if (options.fork)
+		event_save = 1;
+#endif
+	/* Similar cludge for MPI. Only SIGUSR1 is supported for showing
+	   status because the fascist MPI daemons will send us a SIGKILL
+	   seconds after a SIGHUP and there's nothing we can do about it. */
+#ifdef HAVE_MPI
+	if (mpi_p > 1 || getenv("OMPI_COMM_WORLD_SIZE"))
+		event_save = 1;
+#endif
 	event_status = event_pending = 1;
-	signal(signum, sig_handle_status);
+#ifndef SA_RESTART
+	sig_install(sig_handle_status, signum);
+#endif
 }
 
 static void sig_handle_reload(int signum)
 {
 #if OS_FORK
-	if (!event_reload) {
+	if (!event_reload && options.fork) {
 		if (john_main_process)
 			signal_children(signum);
 		else
 			kill(getppid(), signum);
 	}
 #endif
-#ifdef HAVE_MPI
-	/*
-	 * This only works if 'mpirun' is parent, not 'orted'. We
-	 * might overcome that by using MPI_Isend() to the root node
-	 * which can MPI_Bcast() to the others. Though that would
-	 * need to happen elsewehere - this is a signal handler %-)
-	 */
-	if (!event_reload && mpi_p > 1)
-		kill(getppid(), signum);
-#endif
-	event_reload = 1;
-	signal(signum, sig_handle_reload);
+	if (!event_abort)
+		event_reload = 1;
+	/* Avoid loops from signalling ppid. We re-enable this signal
+	   in sig_handle_timer() */
+	signal(signum, SIG_IGN);
 }
 
 static void sig_done(void);
@@ -453,14 +457,14 @@ void sig_init(void)
 
 	atexit(sig_done);
 
-	sig_install_update();
+	sig_install(sig_handle_update, SIGHUP);
 	sig_install_abort();
 	sig_install_timer();
 #ifdef SIGUSR1
-	signal(SIGUSR1, sig_handle_status);
+	sig_install(sig_handle_status, SIGUSR1);
 #endif
 #ifdef SIGUSR2
-	signal(SIGUSR2, sig_handle_reload);
+	sig_install(sig_handle_reload, SIGUSR2);
 #endif
 }
